requirements.txt
argparse==1.2.1
cycler==0.10.0
decorator==4.0.10
glove-python==0.1.0
matplotlib==1.5.3
networkx==1.11
numpy==1.11.2
pyparsing==2.1.10
python-dateutil==2.5.3
pytz==2016.7
scipy==0.18.1
six==1.10.0
wsgiref==0.1.2
